Add CPPUNIT_ASSERT_PRED
I'e added a featurette to our local fork of CPPUnit that allows you to
assert on a predicate (CPPUNIT_ASSERT_PREDICATE) and gives the user a
useful message if the predicate fails.
This allows you to implement things like this:
	typedef std::less< long > Less;
	CPPUNIT_ASSERT_PREDICATE("Update should happen within 10 us", 10, delay,
Less());
(the typedef is necessary because templates and macros don't mix
well.)
I've followed the existing code as a style guide and included Doxygen
comments where appropriate.

This patch was originally developed at Cooper Power Systems EAS (a
division of Cooper Industries) and submitted to the CPPUnit developer's
list with permission from my supervisor (Pierre Martin-Dumont) by
myself. Not getting a response to that submission was the motivator for
forking CPPUnit.

Signed-off-by: Ronald Landheer-Cieslak <rlc@vlinder.ca>