cppunit 1.12.1 needs -lm for fabs() on solaris
I just finished building cppunit 1.12.1 on x86_64-sun-solaris2.10 with
the no-cost Oracle Studio 12.2 compilers.

cppunit needs the fabs() function, and currently it's assuming it will
be available. Solaris has it, but it's in libm (the math library).
That's also where a few other commercial UNIX platforms would likely
have it as well.

The fix is trivial; check for fabs in libm.

With this patch and my previous patch related to the finite() prototype
in ieeefp.h, cppunit 1.12.1 builds in 64 bit mode with the Oracle
compilers and all the expected tests pass.

Signed-off-by: Ronald Landheer-Cieslak <rlc@vlinder.ca>