cppunit 1.12.1 needs ieeefp.h for finite() protype on Solaris
I build cppunit 1.12.1 on x86_64-sun-solaris2.10 with the no-cost Oracle
Studio 12.2 toolchain.

configure for cppunit correct detects that Solaris has finite() (and
it's in libc -- most of Solaris' math-related functions are in libm),
but for reasons I can only guess at, the prototype for finite() is not
in math.h or cmath -- it's in ieeefp.h.

The fix is simple. Have configure.in check for ieeefp.h, and include in
cppunit/portability/FloatingPoint.h if it's present. Patch is attached.
An autoheader & autoconf are needed after this.

Signed-off-by: Ronald Landheer-Cieslak <rlc@vlinder.ca>